Rename logging to logger to avoid import conflict
