+++ siiptool/scripts/siip_sign.py
@@ -24,7 +24,7 @@ from ctypes import c_char, c_uint32, c_uint8, c_uint64, c_uint16, sizeof, ARRAY
 sys.path.insert(0, os.path.dirname(os.path.dirname(os.path.abspath(__file__))))
 from common.siip_constants import VERSION as __version__
 from common.banner import banner
-import common.logging as logging
+import common.logger as logging
 
 logger = logging.getLogger("siip_sign")
